Consolidate auth validation
