+++ jib-core/src/main/java/com/google/cloud/tools/jib/plugins/common/AuthProperty.java
@@ -1,0 +1,13 @@
+package com.google.cloud.tools.jib.plugins.common;
+
+import javax.annotation.Nullable;
+
+/** Holds a username and password property. */
+public interface AuthProperty {
+
+  @Nullable
+  String getUsername();
+
+  @Nullable
+  String getPassword();
+}
