+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildImageMojo.java
@@ -24,6 +24,7 @@ import com.google.cloud.tools.jib.image.ImageFormat;
 import com.google.cloud.tools.jib.image.ImageReference;
 import com.google.cloud.tools.jib.plugins.common.BuildStepsExecutionException;
 import com.google.cloud.tools.jib.plugins.common.BuildStepsRunner;
+import com.google.cloud.tools.jib.plugins.common.ConfigurationPropertyValidator;
 import com.google.cloud.tools.jib.plugins.common.HelpfulSuggestions;
 import com.google.cloud.tools.jib.registry.credentials.RegistryCredentials;
 import com.google.common.annotations.VisibleForTesting;
@@ -83,9 +84,10 @@ public class BuildImageMojo extends JibPluginConfiguration {
     ImageReference targetImage =
         PluginConfigurationProcessor.parseImageReference(getTargetImage(), "to");
     Authorization toAuthorization =
-        PluginConfigurationProcessor.getImageAuth(
+        ConfigurationPropertyValidator.getImageAuth(
             mavenJibLogger,
-            "to",
+            "<to><auth><username>",
+            "<to><auth><password>",
             "jib.to.auth.username",
             "jib.to.auth.password",
             getTargetImageAuth());
